Fixes SQLSTATE[22007]: Invalid datetime format: 1292 Truncated incorrect DOUBLE value: 'GiTd0xvE' (SQL: delete from `has_product` where `holder_id` = 36 and `holder_type` = App\\Models\\User)
This problem is caused by a polymorphic many-to-many pivot table that has string IDs for the foreign pivot key (this is so you can store incrementing IDs and UUIDs together in the pivot table). If the owner of the morphed table (the class that has morphToMany()) has incrementing IDs (integers), Eloquent will create a query comparing the pivot table to that number. Since the pivot table has string foreign keys, SQL will throw this error in strict mode.

This commit attempts to solve this by manually passing all the IDs cast to string.